update TokenUiComponentProvider to display token if condition is met
