+++ Model/Ui/TokenUiComponentProvider.php
@@ -11,13 +11,29 @@
 
 namespace Adyen\Payment\Model\Ui;
 
+use Adyen\Payment\Helper\Data;
+use Adyen\Payment\Helper\Vault;
+use Adyen\Payment\Helper\Recurring;
 use Magento\Vault\Api\Data\PaymentTokenInterface;
 use Magento\Vault\Model\Ui\TokenUiComponentInterface;
 use Magento\Vault\Model\Ui\TokenUiComponentProviderInterface;
 use Magento\Vault\Model\Ui\TokenUiComponentInterfaceFactory;
 
-class TokenUiComponentProvider extends AdyenUiComponentProvider implements TokenUiComponentProviderInterface
+class TokenUiComponentProvider implements TokenUiComponentProviderInterface
 {
+    /**
+     * @param TokenUiComponentInterfaceFactory $componentFactory
+     * @param Data $dataHelper
+     */
+    public function __construct(
+        TokenUiComponentInterfaceFactory $componentFactory,
+        Data $dataHelper,
+        Vault $vaultHelper
+    ) {
+        $this->componentFactory = $componentFactory;
+        $this->dataHelper = $dataHelper;
+        $this->vaultHelper = $vaultHelper;
+    }
 
     /**
      * Get UI component for token
@@ -27,10 +43,20 @@ class TokenUiComponentProvider extends AdyenUiComponentProvider implements Token
      */
     public function getComponentForToken(PaymentTokenInterface $paymentToken): TokenUiComponentInterface
     {
-        $token = $paymentToken->getData()['details'];
-        if(!isset($token) || str_contains($token, 'CardOnFile')) {
-            echo 'hoi';
-        }
-        return $this->getCardComponentForToken($paymentToken);
+        $tokenType = $this->vaultHelper->getAdyenTokenType($paymentToken);
+        $details = json_decode($paymentToken->getTokenDetails() ?: '{}', true);
+        $details['icon'] = $this->dataHelper->getVariantIcon($details['type']);
+        $details['displayToken'] = $tokenType === Recurring::CARD_ON_FILE || !isset($tokenType);
+
+        return $this->componentFactory->create(
+            [
+                'config' => [
+                    'code' => AdyenCcConfigProvider::CC_VAULT_CODE,
+                    TokenUiComponentProviderInterface::COMPONENT_DETAILS => $details,
+                    TokenUiComponentProviderInterface::COMPONENT_PUBLIC_HASH => $paymentToken->getPublicHash()
+                ],
+                'name' => 'Adyen_Payment/js/view/payment/method-renderer/vault'
+            ]
+        );
     }
 }
